+++ src/styles/index.scss
@@ -11,3 +11,6 @@
     import 'relative/path/to/stylesheet.scss';
 
 */
+.task {
+    display: inline-block;
+}
